Version 0.62
- Timezone issue (#75)
- Added some fixes to allow installation on not-supervised env

+++ rootfs/etc/cont-init.d/simplescheduler.sh
@@ -10,7 +10,9 @@ FOLDER=/share/simplescheduler
 
 mkdir -p $FOLDER
 
-cp /data/options.json $FOLDER/options.dat
+if [ -f /data/options.json ]; then
+	cp /data/options.json $FOLDER/options.dat
+fi
 
 # In older version the option file was named option.json
 # If such file exists, delete it
